requirements.txt
Flask==2.2.3
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==8.4.0
opencv-python==4.7.0.72
scikit-learn==1.3.0
numpy==1.25.1
spacy
pypdf2